create observable proxy
