+++ packages/ag-grid/src/lib/types/index.ts
@@ -1,3 +1,2 @@
 export * from './gridConfig';
 export * from './gridController';
-export * from './proxy';
