+++ packages/observable-proxy/.babelrc
@@ -1,0 +1,10 @@
+{
+	"presets": [
+		[
+			"@nrwl/web/babel",
+			{
+				"useBuiltIns": "usage"
+			}
+		]
+	]
+}
