+++ packages/observable-proxy/README.md
@@ -1,0 +1,7 @@
+# observable-proxy
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test observable-proxy` to execute the unit tests via [Jest](https://jestjs.io).
