+++ packages/observable-proxy/jest.config.ts
@@ -1,0 +1,15 @@
+/* eslint-disable */
+export default {
+	displayName: 'observable-proxy',
+	preset: '../../jest.preset.js',
+	globals: {
+		'ts-jest': {
+			tsconfig: '<rootDir>/tsconfig.spec.json',
+		},
+	},
+	transform: {
+		'^.+\\.[tj]s$': 'ts-jest',
+	},
+	moduleFileExtensions: ['ts', 'js', 'html'],
+	coverageDirectory: '../../coverage/packages/observable-proxy',
+};
