+++ packages/observable-proxy/src/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './createObservableProxy';
+export * from './types';
