+++ packages/observable-proxy/src/lib/types/subscriberProxy.ts
@@ -1,12 +1,12 @@
 import { Observable } from 'rxjs';
 
 /**Picks keys that extends the given type */
-export type PickTypeKeys<Obj, Type, T extends keyof Obj = keyof Obj> = {
+type PickTypeKeys<Obj, Type, T extends keyof Obj = keyof Obj> = {
 	[P in keyof Obj]: Obj[P] extends Type ? P : never;
 }[T];
 
 /**Omits all keys of a certain type */
-export type OmitType<T, Type> = Omit<T, PickTypeKeys<T, Type>>;
+type OmitType<T, Type> = Omit<T, PickTypeKeys<T, Type>>;
 
 /**
  * Creates a union of an object and creates a Observable of all properties
