+++ packages/observable-proxy/tsconfig.lib.json
@@ -1,0 +1,10 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"outDir": "../../dist/out-tsc",
+		"declaration": true,
+		"types": []
+	},
+	"include": ["**/*.ts"],
+	"exclude": ["jest.config.ts", "**/*.spec.ts", "**/*.test.ts"]
+}
