+++ tsconfig.base.json
@@ -18,6 +18,7 @@
 		"paths": {
 			"@equinor/filter": ["packages/filter/src/index.ts"],
 			"@equinor/garden": ["packages/garden/src/index.ts"],
+			"@equinor/observable-proxy": ["packages/observable-proxy/src/index.ts"],
 			"@equinor/powerbi": ["packages/power-bi/src/index.ts"],
 			"@equinor/sidesheet": ["packages/sidesheet/src/index.ts"],
 			"@equinor/status-bar": ["packages/status-bar/src/index.ts"],
